4_projeto/ecommerce/requirements.txt
asgiref==3.6.0
astroid==2.15.4
autopep8==2.0.2
dill==0.3.6
Django==4.2.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.5.0
platformdirs==3.5.0
pycodestyle==2.10.0
pylint==2.17.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
wrapt==1.15.0
